moved paths to logging module and made configurable path
